Add cached gRPC parity fixture for Zebra-backed comparison

+++ qa/rpc-tests/test_framework/util.py
@@ -616,10 +616,9 @@ def initialize_chain(test_dir, num_nodes, cachedir, cache_behavior='current'):
 
             # Copy in per-node wallet data
             wallet_tgz_filename = os.path.join(cache_path, "node"+str(i)+"_wallet.tar.gz")
-            if not os.path.exists(wallet_tgz_filename):
-                raise Exception('Wallet cache missing for cache behavior %s, node %d' % (cache_behavior, i))
-            with tarfile.open(wallet_tgz_filename, "r:gz") as wallet_tgz_file:
-                tarfile_extractall(wallet_tgz_file, os.path.join(to_dir, "wallet.dat"))
+            if os.path.exists(wallet_tgz_filename):
+                with tarfile.open(wallet_tgz_filename, "r:gz") as wallet_tgz_file:
+                    tarfile_extractall(wallet_tgz_file, os.path.join(to_dir, "wallet.dat"))
 
             # Copy in per-node wallet config and update zcash.conf to set the
             # clock offsets correctly.
@@ -686,23 +685,27 @@ def persist_node_caches(tmpdir, cache_behavior, num_nodes):
         node_path = os.path.join(tmpdir, 'node' + str(i), 'regtest')
 
         # Clean up the files that we don't want to persist
-        os.remove(os.path.join(node_path, 'debug.log'))
-        os.remove(os.path.join(node_path, 'db.log'))
-        os.remove(os.path.join(node_path, 'peers.dat'))
+        for filename in ('debug.log', 'db.log', 'peers.dat'):
+            path = os.path.join(node_path, filename)
+            if os.path.exists(path):
+                os.remove(path)
 
         # Persist the wallet file for the node to the cache
+        wallet_path = os.path.join(node_path, 'wallet.dat')
         wallet_tgz_filename = os.path.join(cache_path, 'node' + str(i) + '_wallet.tar.gz')
-        with tarfile.open(wallet_tgz_filename, "w:gz") as wallet_tgz_file:
-            wallet_tgz_file.add(os.path.join(node_path, 'wallet.dat'), arcname="")
+        if os.path.exists(wallet_path):
+            with tarfile.open(wallet_tgz_filename, "w:gz") as wallet_tgz_file:
+                wallet_tgz_file.add(wallet_path, arcname="")
 
         # Persist the chain data and cache config just once; it will be reused
         # for all of the nodes when loading from the cache.
         if i == 0:
             # Move the wallet.dat file out of the way so that it doesn't
             # pollute the chain cache tarfile
-            shutil.move(
-                    os.path.join(node_path, 'wallet.dat'),
-                    os.path.join(tmpdir, 'wallet.dat.0'))
+            if os.path.exists(wallet_path):
+                shutil.move(
+                        wallet_path,
+                        os.path.join(tmpdir, 'wallet.dat.0'))
 
             # Store the current time so that we can correctly set the clock
             # offset when restoring from the cache.
@@ -718,9 +721,11 @@ def persist_node_caches(tmpdir, cache_behavior, num_nodes):
                 chain_cache_file.add(node_path, arcname="")
 
             # Move the wallet file back into place
-            shutil.move(
-                    os.path.join(tmpdir, 'wallet.dat.0'),
-                    os.path.join(node_path, 'wallet.dat'))
+            wallet_tmp = os.path.join(tmpdir, 'wallet.dat.0')
+            if os.path.exists(wallet_tmp):
+                shutil.move(
+                        wallet_tmp,
+                        wallet_path)
 
 
 def _rpchost_to_args(rpchost):
@@ -1384,7 +1389,8 @@ def update_zainod_conf(datadir, rpc_port, indexer_port, zaino_rpc_port, zaino_gr
         json_rpc_listen_address='127.0.0.1:'+str(zaino_rpc_port),
         grpc_listen_address='127.0.0.1:'+str(zaino_grpc_port),
         validator_grpc_listen_address='127.0.0.1:'+str(indexer_port),
-        validator_jsonrpc_listen_address='127.0.0.1:'+str(rpc_port)
+        validator_jsonrpc_listen_address='127.0.0.1:'+str(rpc_port),
+        storage_database_path=os.path.join(datadir, 'db'),
     )
 
     config_file = zaino_config.update(config_file)
@@ -2040,12 +2046,12 @@ def zcashd_rpc_url(i):
     return "http://%s:%s@127.0.0.1:%d" % (ZCASHD_RPC_USER, ZCASHD_RPC_PASSWORD, zcashd_rpc_port(i))
 
 
-def write_zcash_conf(datadir, node_rpc_port, node_p2p_port, miner_address=None):
+def write_zcash_conf(datadir, node_rpc_port, node_p2p_port, miner_address=None, activation_heights=None):
     """Write a zcash.conf for a standalone regtest zcashd node.
 
-    All network upgrades are activated from block 1.  Heartwood (ZIP 213)
-    enables shielded coinbase; NU5 enables Orchard.  No addnode is written
-    because blocks are pushed to zebrad via submitblock rather than P2P.
+    By default all network upgrades are activated from block 1 to preserve the
+    historical behavior of the standalone helper. Tests can override the
+    activation heights to match a validator's regtest configuration.
 
     miner_address, if given, sets the coinbase recipient for `generate` calls.
     setmineraddress has been removed from this zcashd version, so the address
@@ -2062,22 +2068,29 @@ def write_zcash_conf(datadir, node_rpc_port, node_p2p_port, miner_address=None):
         f.write("rpcpassword=%s\n" % ZCASHD_RPC_PASSWORD)
         if miner_address is not None:
             f.write("mineraddress=%s\n" % miner_address)
-        # Activate all upgrades from block 1 to match zebrad Regtest defaults.
-        f.write("nuparams=5ba81b19:1\n")   # Overwinter
-        f.write("nuparams=76b809bb:1\n")   # Sapling
-        f.write("nuparams=2bb40e60:1\n")   # Blossom
-        f.write("nuparams=f5b9230b:1\n")   # Heartwood (ZIP 213: shielded coinbase)
-        f.write("nuparams=e9ff75a6:1\n")   # Canopy
-        f.write("nuparams=c2d6d0b4:1\n")   # NU5 (Orchard)
-        f.write("nuparams=c8e71055:1\n")   # NU6
+            f.write("minetolocalwallet=0\n")
+        if activation_heights is None:
+            activation_heights = {
+                '5ba81b19': 1,  # Overwinter
+                '76b809bb': 1,  # Sapling
+                '2bb40e60': 1,  # Blossom
+                'f5b9230b': 1,  # Heartwood
+                'e9ff75a6': 1,  # Canopy
+                'c2d6d0b4': 1,  # NU5
+                'c8e71055': 1,  # NU6
+            }
+        for branch_id, height in activation_heights.items():
+            f.write("nuparams=%s:%d\n" % (branch_id, height))
         # Re-enable the deprecated getnewaddress RPC used to obtain a t-address
         # for transparent coinbase mining.
         f.write("allowdeprecated=getnewaddress\n")
+        f.write("allowdeprecated=z_getnewaddress\n")
+        f.write("regtestwalletsetbestchaineveryblock=1\n")
         f.write("i-am-aware-zcashd-will-be-replaced-by-zebrad-and-zallet-in-2025=1\n")
     return conf_path
 
 
-def start_zcashd_node(i, dirname, miner_address=None, binary=None, stderr=None):
+def start_zcashd_node(i, dirname, miner_address=None, activation_heights=None, binary=None, stderr=None):
     """Start a standalone regtest zcashd node and return an RPC proxy."""
     if binary is None:
         binary = zcashd_node_binary()
@@ -2087,6 +2100,7 @@ def start_zcashd_node(i, dirname, miner_address=None, binary=None, stderr=None):
         zcashd_rpc_port(i),
         zcashd_p2p_port(i),
         miner_address=miner_address,
+        activation_heights=activation_heights,
     )
     args = [binary, "-conf=" + conf, "-datadir=" + datadir]
     zcashd_node_processes[i] = subprocess.Popen(args, stderr=stderr)
